Added DistributionBayesianNetwork
This class allows to create and manipulate general continuous bayesian networks given a DAG and a set of distributions for each node. For nodes without parents, the distribution is its marginal distribution, for the other nodes it is a way to define the conditional distribution of the node given its parents.